# Refactor triggers and greatly simplify action change tracking