--- a/vertex/layer4/action-runner.test.ts
+++ b/vertex/layer4/action-runner.test.ts
@@ -116,11 +116,7 @@ group("action runner", () => {
                     tx.run("CREATE (x:SomeNode) RETURN x", {})
                 ),
                 undefined,
-                "Error executing triggers {trackActionChanges=Unable to complete transaction.}",
-                // Should be: "every data write transaction should be associated with one Action"
-                // I tried every fucking way to get the trackActionChanges to produce this error like it used to, but
-                // Neo4j just doesn't like to allow raising exceptions and writing to relationship properties in a
-                // trigger transaction :-/  (Uncommenting the "SET modRel += ..." line will restore the custom errors)
+                "every data write transaction should be associated with one Action",
             );
         });
     });
@@ -150,11 +146,7 @@ group("action runner", () => {
         await assertThrowsAsync(
             () => testGraph.runAsSystem( CreateCeresAction({markAsModified: false}) ),
             undefined,
-            "Error executing triggers {trackActionChanges=Unable to complete transaction.}",
-            // Should be: "A :AstroBody node was modified by this CreateCeres1 action (created) but not explicitly marked as modified by the Action.",
-            // I tried every fucking way to get the trackActionChanges to produce this error like it used to, but
-            // Neo4j just doesn't like to allow raising exceptions and writing to relationship properties in a
-            // trigger transaction :-/  (Uncommenting the "SET modRel += ..." line will restore the custom errors)
+            "A AstroBody node was modified by this action but not explicitly marked as modified by the Action.",
         );
 
         // Then it should work if it does mark the node as modified:
@@ -174,11 +166,7 @@ group("action runner", () => {
         await assertThrowsAsync(
             () => testGraph.runAsSystem(GenericCypherAction({cypher, modifiedNodes: []})),
             undefined,
-            "Error executing triggers {trackActionChanges=Unable to complete transaction.}",
-            // Should be: "A :AstroBody node was modified by this GenericCypherAction action (newProp:mass) but not explicitly marked as modified by the Action.",
-            // I tried every fucking way to get the trackActionChanges to produce this error like it used to, but
-            // Neo4j just doesn't like to allow raising exceptions and writing to relationship properties in a
-            // trigger transaction :-/  (Uncommenting the "SET modRel += ..." line will restore the custom errors)
+            "A AstroBody node was modified by this action but not explicitly marked as modified by the Action.",
         );
         // Then it should work if it does mark the node as modified:
         await testGraph.runAsSystem(GenericCypherAction({cypher, modifiedNodes: [id]}));