* Remove API for tracking action changes in detail (should be plugin)
   - most apps using this framework probably won't want/need to have a complete record of every action change, so this behavior should be moved to a plugin. This simplifies the core triggers significantly.
* Refactor triggers to simplify them and clearly separate **validation** triggers (that can throw exceptions) from **modification** triggers that update data.
   - this results in *much* clearer error messages being reported to the user and makes debugging way easier

This is a **major breaking change** and it edits some existing migrations.

## Benchmark 2022-07-02

Before: OpenAlex levels 0-2:

```
Run 1
level 0: 14,382ms
level 1: 24,753ms
level 2: 2,801,685ms
overall: 2,840,820ms
```

Before: OpenAlex levels 0-1:

```
Run 1
level 0: 14,353ms
level 1: 25,286ms
overall: 39,639ms
```

After: OpenAlex levels 0-1:

```
Run 1
level 0: 14,353ms
level 1: 25,254ms
overall: 39,608ms
```

Seems like there is no discernable performance impact.

Also the actual "run tests" part of the test suite ran faster on this branch than main (36 vs 40 s).